% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyseMRIO-package.R
\docType{package}
\name{AnalyseMRIO-package}
\alias{AnalyseMRIO}
\alias{AnalyseMRIO-package}
\title{AnalyseMRIO: What the Package Does (One Line, Title Case)}
\description{
Reprise des fonctions du projet sous la forme d'un package R.
}
\author{
\strong{Maintainer}: Tanguy Richard \email{lorenzitanguy@gmail.com}

}
\keyword{internal}
