% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_A17-data.R
\docType{data}
\name{BASE_A17}
\alias{BASE_A17}
\title{MRIO A17}
\format{
Le MRIO A17 sous la forme d'un dataframe dont chaque ligne représente une case
du document MRIO
year : année de mesure du MRIO
Lig_Country : Pays correspondant à la ligne du MRIO
Col_Country : Pays correspondant à la colonne du MRIO
Lig_Indus : Industrie correspondant à la ligne du MRIO
Col_Indus : Industrie correspondant à la colonne du MRIO
value : Valeur de la case correspondant aux coordonnées précisées dans les autres colonnes
MRIO : MRIO d'origine
}
\source{
Alexandre Bourgeois
}
\usage{
data(BASE_A17)
}
\description{
MRIO A17
}
\examples{
\dontrun{
data(BASE_A17)
}
}
\keyword{datasets}
