% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_ECOLE-data.R
\docType{data}
\name{BASE_ECOLE}
\alias{BASE_ECOLE}
\title{MRIO Ecole}
\format{
Le MRIO école sous la forme d'un dataframe dont chaque ligne représente une case
du document MRIO
Soit un dataframe à 59026 lignes et 7 variables:
year : année de mesure du MRIO
Lig_Country : Pays correspondant à la ligne du MRIO
Col_Country : Pays correspondant à la colonne du MRIO
Lig_Indus : Industrie correspondant à la ligne du MRIO
Col_Indus : Industrie correspondant à la colonne du MRIO
value : Valeur de la case correspondant aux coordonnées précisées dans les autres colonnes
MRIO : MRIO d'origine
}
\source{
Alexandre Bourgeois
}
\usage{
data(BASE_ECOLE)
}
\description{
MRIO Ecole
}
\examples{
\dontrun{
data(BASE_ECOLE)
}
}
\keyword{datasets}
