% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralite.R
\name{centralite}
\alias{centralite}
\title{centralite}
\usage{
centralite(SPA_annee, node, by = "count")
}
\arguments{
\item{SPA_annee}{data.table issue de SPA_opt2}

\item{node}{nom du pays_secteur}

\item{by}{Soit "count" pour compter le nombre de chemin, soit "value" pour la contribution total}
}
\value{
la valeur correspondante
}
\description{
calcul le nombre de passage par un noeud (pays_secteur) pour les chemins de longueur 3 ou plus
C'est une mesure de betweenness
}
\examples{
\dontrun{

data(SPA_ECOLE)

# Pour dénombrer le nombre de chemin passant par le noeud en intermédiaire
centralite(SPA_ECOLE$don1990, "CHN_AGR_INDU")

# Pour comptabiliser la valeur totale des chemin passant par le noeud en intermédiaire
centralite(SPA_ECOLE$don1990, "CHN_AGR_INDU", by="value")

}
}
