% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralite_annee.R
\name{centralite_annee}
\alias{centralite_annee}
\title{centralite_annee}
\usage{
centralite_annee(SPA_annee, by = "count")
}
\arguments{
\item{SPA_annee}{data.table issue de SPA_opt2}

\item{by}{Soit "count" pour compter le nombre de chemin, soit "value" pour la contribution total}
}
\value{
un data.frame avec une colonne noeud et une colonne correspondant à la valeur de l'indice
}
\description{
Récupère un tableau de d'indice de betweenness (compte ou valeur total)
pour toute les noeuds pays/secteur
}
\examples{
\dontrun{

data(SPA_ECOLE)

# Pour dénombrer le nombre de chemin passant par le noeud en intermédiaire
centralite_annee(SPA_ECOLE$don1990)

# Pour comptabiliser la valeur totale des chemin passant par le noeud en intermédiaire
centralite_annee(SPA_ECOLE$don1990, by = "value")

}
}
