% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralite_totale.R
\name{centralite_totale}
\alias{centralite_totale}
\title{centralite_totale}
\usage{
centralite_totale(SPA, by = "count")
}
\arguments{
\item{SPA}{liste de data.table issue de SPA_opt2}

\item{by}{Soit "count" pour compter le nombre de chemin, soit "value" pour la contribution total}
}
\value{
un data.frame avec une colonne noeud et une colonne par année correspondant à la valeur de l'indice
}
\description{
centralite_totale
}
\examples{
\dontrun{

data(SPA_ECOLE)

# Pour dénombrer le nombre de chemin passant par le noeud en intermédiaire
centralite_totale(SPA_ECOLE)

# Pour comptabiliser la valeur totale des chemin passant par le noeud en intermédiaire
centralite_totale(SPA_ECOLE, by = "value")

}
}
