% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CFPcalculationRCPP}
\alias{CFPcalculationRCPP}
\title{Simple matrix multiplication}
\usage{
CFPcalculationRCPP(ef, L, FD)
}
\arguments{
\item{ef}{Matrix}

\item{L}{Matrix}

\item{FD}{Matrix}
}
\value{
Product of matrices
}
\description{
Simple matrix multiplication
}
\examples{
A <- matrix(1:9, 3, 3)
B <- matrix(11:19, 3, 3)
C <- matrix(4:12, 3, 3)
CFPcalculationRCPP(A, B, C)
}
