% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoECOLEouA17.R
\name{CompoECOLEouA17}
\alias{CompoECOLEouA17}
\title{CompoECOLEouA17}
\usage{
CompoECOLEouA17(
  dt_MRIO,
  typeCompo,
  date = 9999,
  country = "ALL",
  OptTab = FALSE,
  OptMRIO1014 = "FIGARO",
  OptMRIO2000 = "WIOD"
)
}
\arguments{
\item{dt_MRIO}{Data frame soit Ecole, soit A17}

\item{typeCompo}{Une composante parmi :
\itemize{
\item "CI"                     : Consommation intermédiaire
\item "DF"                     : Demande finale
\item "PROD"                   : Production
\item "VA"                     : Valeur ajouté
\item "A"                      : Matrice des coefficients techniques (Leontief)
\item "B"                      :
\item "CI_PR"                  :
\item "CI_BR"                  :
\item "DF_TOT"                 :
\item "L"                      : Inverse de Leontief
\item "invB"                   :
\item "OptFullOptions"         : Liste de resultat
\item "OptFullOptionsBonus"    : Liste de l'ensemble des resultats
}}

\item{date}{Année parmi celle présente dans le data frame}

\item{country}{Pays parmi ceux du data frame}

\item{OptTab}{Option pour avoir un resultat sous forme de tableau (défault : FALSE)}

\item{OptMRIO1014}{Option de choix de base sur la période 2010-2014 : par défaut c’est FIGARO qui est retenu}

\item{OptMRIO2000}{Option de choix de base pour l’année 2000 : par défaut c’est WIOD qui est retenu}
}
\value{
Un data frame correspondant à la selection
ou pour les 2 options : "OptFullOptions" et "OptFullOptionsBonus"  une liste de data frame
}
\description{
Sélection d'une composante du mrio ECOLE ou A17 (attention priorité par défaut : Figaro sur 2010-2014)
}
\examples{
\dontrun{

data(BASE_ECOLE)
MRIO_1990_ALL<-CompoECOLEouA17(BASE_ECOLE,"OptFullOptionsBonus",date=1990)
summary(MRIO_1990_ALL$L)

}
}
