% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consomation.R
\name{consomation}
\alias{consomation}
\title{consomation}
\usage{
consomation(SPA_annee, pays, by = "secteurs", listepays = NA)
}
\arguments{
\item{SPA_annee}{data.table issue de SPA_opt2}

\item{pays}{Code pays présent dans les données}

\item{by}{"secteurs" ou "pays" selon l'aggrégation souhaité}

\item{listepays}{liste de code pays des pays nous interessant pour l'analyse (nécessaire pour by ="pays")}
}
\value{
data.table à 2 colonnes correspondant au nom du pays ou secteurs et la valeur de la contribution totale
}
\description{
tableau de l'impacte des consomation intermédiaire par pays ou secteur
}
\examples{
\dontrun{

data(SPA_ECOLE)

don1990 <- SPA_ECOLE$don1990

tt <- consomation(don1990,"FRA")
View(tt)

tt2 <- consomation(don1990,"FRA", by = "pays", listepays = c("FRA","CHN","ROW","UE_OTHERS","USA"))
View(tt2)

}
}
