% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contribution_SubTree.R
\name{Contribution_SubTree}
\alias{Contribution_SubTree}
\title{Contribution_SubTree}
\usage{
Contribution_SubTree(f, A, Y, Z)
}
\arguments{
\item{f}{vecteur des contributions par pays_secteur par unité de production}

\item{A}{matrice de Leontief}

\item{Y}{Demande Finale}

\item{Z}{Noeud (sous la forme d'un vecteur de pays_secteur)}
}
\value{
La valeur de la contribution finale
}
\description{
Calcule la contribution du sous arbre issue du noeud
}
\examples{
\dontrun{

data(BASE_ECOLE)

# Pour avoir la matrice de Leontief :
MRIO_1990_ALL<-CompoECOLEouA17(BASE_ECOLE,"OptFullOptionsBonus",date=1990)
A <- MRIO_1990_ALL[["A_tab"]][,-1] # On enlève la colonne année
A <- MatCol(A,"Lig_Country","Lig_Indus")

# Pour aller chercher Y (demande finale)
Y <- MRIO_1990_ALL[["DF_TOT"]]
Y <- Y[,-c(1,2)] # prendre uniquement la colonne d'interet
Y <- MatCol(Y,"Lig_Country","Lig_Indus")

# Pour aller chercher VA/PROD
VA <- MRIO_1990_ALL[["VA"]]
PROD <- MRIO_1990_ALL[["PROD"]]
VA <- VA[,-c(1,2)] # prendre uniquement la colonne d'interet
PROD <- PROD[,-c(1,2)]
VA <- MatCol(VA,"Col_Country","Col_Indus")
PROD <- MatCol(PROD,"Lig_Country","Lig_Indus")
VA <- VA/PROD

# On donnes un parcours
Z <- c( "CHN_ENRJ", "CHN_AGR_INDU", "UE_OTHERS_AGR_INDU")

# Contribution du sous arbre
Contribution_SubTree(VA,A,Y,Z)
}
}
