% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courbes_consos.R
\name{courbes_consos}
\alias{courbes_consos}
\title{courbes_consos}
\usage{
courbes_consos(
  SPA,
  pays,
  by = "secteurs",
  listepays = NA,
  prop = TRUE,
  liss = FALSE
)
}
\arguments{
\item{SPA}{liste de data.table issue de SPA_opt2}

\item{pays}{Code pays présent dans les données}

\item{by}{"secteurs" ou "pays" selon l'aggrégation souhaité}

\item{listepays}{liste de code pays des pays nous interessant pour l'analyse (nécessaire pour by ="pays")}

\item{prop}{si TRUE, affiche la proportion du total par année, si FALSE, affiche les valeurs brutes}

\item{liss}{ajouter un lissage linéaire sur les courbes}
}
\value{
un graphique ggplot2 correspondant à la demande
}
\description{
Courbe de l'évolution de l'impacte de pays ou secteurs dans les productions intermédiaire d'un pays
}
\examples{
\dontrun{

data(SPA_ECOLE)

courbes_consos(SPA_ECOLE,"FRA", by = "pays",
 listepays = c("FRA","CHN","ROW","UE_OTHERS","USA"), liss = TRUE)

courbes_consos(SPA_ECOLE,"FRA", by = "secteurs",prop = FALSE)

}
}
