% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonct_temp.R
\name{fonct_temp}
\alias{fonct_temp}
\title{Construct_Tree}
\usage{
fonct_temp(f, A, Y, Tmax, tol, Z = "Tout", L, Ftot)
}
\arguments{
\item{f}{vecteur des contributions par pays_secteur par unité de production}

\item{A}{matrice de Leontief}

\item{Y}{Demande Finale}

\item{Tmax}{Profondeur maximale autorisée}

\item{tol}{Valeur limite}

\item{Z}{Noeud actuel}

\item{L}{inverse de Leontief}

\item{Ftot}{effet tottal}
}
\value{
Un arbre sous la forme d'une liste de 3 éléments :
\itemize{
\item Z vecteur étiquettant le noeud
\item Node Valeur de la contribution du noeud
\item Next Liste étiqueté par les états de :
- sous arbres dépendants du noeud sous la même forme
- NULL s'il n'y a pas de sous arbre
}
}
\description{
Construit récursivement un arbre récapitulant les effets à chaques étapes.
}
\examples{
\dontrun{

data(BASE_ECOLE)

# Pour avoir la matrice de Leontief :
MRIO_1990_ALL<-CompoECOLEouA17(BASE_ECOLE,"OptFullOptionsBonus",date=1990)
A <- MRIO_1990_ALL[["A_tab"]][,-1] # On enlève la colonne année
A <- MatCol(A,"Lig_Country","Lig_Indus")

# Pour aller chercher Y (demande finale)
Y <- MRIO_1990_ALL[["DF_TOT"]]
Y <- Y[,-c(1,2)] # prendre uniquement la colonne d'interet
Y <- MatCol(Y,"Lig_Country","Lig_Indus")

# Pour aller chercher VA/PROD
VA <- MRIO_1990_ALL[["VA"]]
PROD <- MRIO_1990_ALL[["PROD"]]
VA <- VA[,-c(1,2)] # prendre uniquement la colonne d'interet
PROD <- PROD[,-c(1,2)]
VA <- MatCol(VA,"Col_Country","Col_Indus")
PROD <- MatCol(PROD,"Lig_Country","Lig_Indus")
VA <- VA/PROD


# On donnes un parcours
Z <- c( "CHN_ENRJ")

Calcul de paramètre
L <- inversion_rcpp3(diag(ncol(A))-A)
Ftot <- Mult2_rcpp3(t(VA),L)

# Contribution du sous arbre
test <- fonct_temp(VA,A,Y,5,0.5,Z,L,Ftot)

test2 <- fonct_temp(VA,A,Y,5,0.5,L,Ftot)

}
}
