% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GereInfNA.R
\name{GereInfNA}
\alias{GereInfNA}
\title{GereInfNA}
\usage{
GereInfNA(df, impute = 0)
}
\arguments{
\item{df}{un data frame}

\item{impute}{la valeur par laquel les remplacé}
}
\value{
Un data frame avec les NA et Inf remplacé par impute
}
\description{
Fonction pour convertir les infinity et les NA en une autre valeur
}
\examples{
\dontrun{

Age <- c(23, 41, Inf, 58, 26)
df <- data.frame(Age)
GereInfNA(df)

}

}
