% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_sankey.R
\name{graph_sankey}
\alias{graph_sankey}
\title{graph_sankey}
\usage{
graph_sankey(SPA_annee, data_chemin, liste_debut, liste_fin, by = "count")
}
\arguments{
\item{SPA_annee}{data.table issue de SPA_opt2}

\item{data_chemin}{data.frame issu de meilleur_chemin}

\item{liste_debut}{liste de noeuds pour la couche de consomation finale}

\item{liste_fin}{liste de noeuds pour la couche de production intermédiaire lié au chemin}

\item{by}{soit "count" pour compter le nombre de chemin, soit "value" pour la contribution total}
}
\value{
un graphique de sankey issu du package networkD3
}
\description{
Produit un graphique de Sankey représentant les chemins d'une longueur fixe donnée par les paramètres choisis
}
