% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inversion_rcpp3}
\alias{inversion_rcpp3}
\title{Inversion de matrice}
\usage{
inversion_rcpp3(m1)
}
\arguments{
\item{m1}{Matrix}
}
\value{
Inverse de la matrice
}
\description{
Inversion de matrice
}
\examples{
A <- diag(2,3)
inversion_rcpp3(A)
}
