% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meilleur_chemin.R
\name{meilleur_chemin}
\alias{meilleur_chemin}
\title{meilleur_chemin}
\usage{
meilleur_chemin(SPA_annee, lon = 3, by = "count")
}
\arguments{
\item{SPA_annee}{data.table issue de SPA_opt2}

\item{lon}{longeur du chemin 3 ou 4}

\item{by}{Soit "count" pour compter le nombre de chemin, soit "value" pour la contribution total}
}
\description{
determine les meilleurs chemins entre noeuds (pays_secteur)
}
\examples{
\dontrun{

data(SPA_ECOLE)

# Pour dénombrer le nombre de chemin passant par le noeud en intermédiaire
meilleur_chemin(SPA_ECOLE$don1990, lon = 3)

# Pour comptabiliser la valeur totale des chemin passant par le noeud en intermédiaire
meilleur_chemin(SPA_ECOLE$don1990, lon = 4, by = "value")

}
}
