% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mult2_rcpp3}
\alias{Mult2_rcpp3}
\title{Simple matrix multiplication}
\usage{
Mult2_rcpp3(matA, matB)
}
\arguments{
\item{matA}{Matrix}

\item{matB}{Matrix}
}
\value{
Product of matrices
}
\description{
Simple matrix multiplication
}
\examples{
A <- matrix(1:9, 3, 3)
B <- matrix(11:19, 3, 3)
Mult2_rcpp3(A, B)
}
