% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restructure_tab.R
\name{restructure_tab}
\alias{restructure_tab}
\title{restructure_tab}
\usage{
restructure_tab(tab_centre, listepays)
}
\arguments{
\item{tab_centre}{une table d'indice de betwenness par ans}

\item{listepays}{la liste des pays présents}
}
\value{
une table à 4 colonnes pays, secteur, valeur et annee
}
\description{
passe une table d'indice de betweenness de wide à long
}
\examples{
\dontrun{

data(SPA)

toute_centre <- centralite_totale(SPA_ECOLE)

pays <- c("CHN","FRA","USA","UE_OTHERS","ROW")

toute_centre_long <- restructure_tab(toute_centre, pays)

}
}
