% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPA_ECOLE-data.R
\docType{data}
\name{SPA_ECOLE}
\alias{SPA_ECOLE}
\title{SPA sur la base école entre 1990 et 2020}
\format{
Une liste de 31 éléments:
chacun nommé "don{année}" correspondant au data.table issue de l'application
d'une SPA de paramêtre tol=2 et Tmax = 10 appliqu" sur l'année corrspondante
}
\source{
Alexandre Bourgeois
}
\usage{
data(SPA_ECOLE)
}
\description{
SPA sur la base école entre 1990 et 2020
}
\examples{
\dontrun{
data(SPA_ECOLE)
}
}
\keyword{datasets}
