% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPA_tt3.R
\name{SPA_tt3}
\alias{SPA_tt3}
\title{SPA_tt3}
\usage{
SPA_tt3(f, A, Y, Tmax, tol, L, Ftot, Z = "Tout", cc = 1)
}
\arguments{
\item{f}{vecteur des contributions par pays_secteur par unité de production}

\item{A}{matrice de Leontief}

\item{Y}{Demande Finale}

\item{Tmax}{Profondeur maximale autorisée}

\item{tol}{Valeur limite}

\item{L}{inverse de Leontief}

\item{Ftot}{effet total}

\item{Z}{Noeud actuel}

\item{cc}{coefficient intemédiaire}
}
\value{
Un data frame à 2 colonnes:
\itemize{
\item une avec le nom du noeud
\item une avec la contribution
}
}
\description{
SPA_tt3
}
\examples{
\dontrun{

data(BASE_ECOLE)

# Pour avoir la matrice de Leontief :
MRIO_1990_ALL<-CompoECOLEouA17(BASE_ECOLE,"OptFullOptionsBonus",date=1990)
A <- MRIO_1990_ALL[["A_tab"]][,-1] # On enlève la colonne année
A <- MatCol(A,"Lig_Country","Lig_Indus")

# Pour aller chercher Y (demande finale)
Y <- MRIO_1990_ALL[["DF_TOT"]]
Y <- Y[,-c(1,2)] # prendre uniquement la colonne d'interet
Y <- MatCol(Y,"Lig_Country","Lig_Indus")

# Pour aller chercher VA/PROD
VA <- MRIO_1990_ALL[["VA"]]
PROD <- MRIO_1990_ALL[["PROD"]]
VA <- VA[,-c(1,2)] # prendre uniquement la colonne d'interet
PROD <- PROD[,-c(1,2)]
VA <- MatCol(VA,"Col_Country","Col_Indus")
PROD <- MatCol(PROD,"Lig_Country","Lig_Indus")
VA <- VA/PROD


# On donnes un parcours
Z <- c( "CHN_ENRJ")
# On donne la valeur du coef associé au chemin
cc <- Y[Z,]

#Calcul de paramètre
L <- inversion_rcpp3(diag(ncol(A))-A)
Ftot <- Mult2_rcpp3(t(VA),L)

# Contribution du sous arbre
test <- SPA_tt3(VA,A,Y,5,0.5,L,Ftot,Z, cc)

test2 <- SPA_tt3(VA,A,Y,5,0.5,L,Ftot)

}
}
