% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variation.R
\name{Variation}
\alias{Variation}
\title{Variation}
\usage{
Variation(chemin, don)
}
\arguments{
\item{chemin}{nom du chemin}

\item{don}{liste de data.table issue de SPA_opt2}
}
\value{
un tableau avec le rang du chemnin par data.table
}
\description{
Determine les rangs d'un chemin dans une liste de resultat de SPA
}
\examples{
\dontrun{

data(SPA_ECOLE)

Variation("CHN_AGR_INDU",SPA_ECOLE)

}
}
