% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectDF.R
\name{vectDF}
\alias{vectDF}
\title{vectDF}
\usage{
vectDF(vect)
}
\arguments{
\item{vect}{Un vecteur}
}
\value{
un vecteur data frame (colonne)
}
\description{
Fonction de transformation en vecteur dataframe
}
\examples{
\dontrun{

v <- c(1,2,3,4)
vectDF(v)

}

}
