% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct_Tree.R
\name{Construct_Tree}
\alias{Construct_Tree}
\title{Construct_Tree}
\usage{
Construct_Tree(f, A, Y, Tmax, tol, Z = "Tout")
}
\arguments{
\item{f}{vecteur des contributions par pays_secteur par unité de production}

\item{A}{matrice de Leontief}

\item{Y}{Demande Finale}

\item{Tmax}{Profondeur maximale autorisée}

\item{tol}{Valeur limite}

\item{Z}{Noeud actuel}
}
\value{
Un arbre sous la forme d'une liste de 3 éléments :
\itemize{
\item Z vecteur étiquettant le noeud
\item Node Valeur de la contribution du noeud
\item Next Liste étiqueté par les états de :
- sous arbres dépendants du noeud sous la même forme
- NULL s'il n'y a pas de sous arbre
}
}
\description{
Construit récursivement un arbre récapitulant les effets à chaques étapes.
}
\examples{
\dontrun{

data(BASE_ECOLE)

# Pour avoir la matrice de Leontief :
MRIO_1990_ALL<-CompoECOLEouA17(BASE_ECOLE,"OptFullOptionsBonus",date=1990)
A <- MRIO_1990_ALL[["A_tab"]][,-1] # On enlève la colonne année
A <- MatCol(A,"Lig_Country","Lig_Indus")

# Pour aller chercher Y (demande finale)
Y <- MRIO_1990_ALL[["DF_TOT"]]
Y <- Y[,-c(1,2)] # prendre uniquement la colonne d'interet
Y <- MatCol(Y,"Lig_Country","Lig_Indus")

# Pour aller chercher VA
VA <- MRIO_1990_ALL[["VA"]]
VA <- VA[,-c(1,2)] # prendre uniquement la colonne d'interet
VA <- MatCol(VA,"Col_Country","Col_Indus")


# On donnes un parcours
Z <- c( "CHN_ENRJ")

# Contribution du sous arbre
test <- Construct_Tree(VA,A,Y,5,4000,Z)

test2 <- Construct_Tree(VA,A,Y,5,4000)

}
}
