% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evolution.R
\name{Evolution}
\alias{Evolution}
\title{Evolution}
\usage{
Evolution(SPA, num)
}
\arguments{
\item{SPA}{liste de data.table issue de SPA_opt2}

\item{num}{nombre de chemins à conserver}
}
\value{
un data frame de l'évolution des rangs
}
\description{
Evolution
}
\examples{
\dontrun{

data(SPA_ECOLE)

t1 <- Evolution(SPA_ECOLE,100)
View(t1)

}
}
